KS-2-1 refactoring

+++ src/main/kotlin/lesson_2/lesson2_task1.kt
@@ -2,10 +2,10 @@ package ru.fedorova.sprint.lesson_2
 
 fun main() {
 
-    val countStudents: Float = 4.0f
+    val countStudents: Int = 4
     val sumPoints: Int = 3 + 4 + 3 + 5
 
-    val averagePoints: Float = sumPoints / countStudents
+    val averagePoints: Float = sumPoints.toFloat() / countStudents
 
     println(averagePoints)
